styling changes and moved header to components
